### 2024-08-26 #7145 @editorialbot
[PRE REVIEW]: Teaspoon: A Python Package for Topological Signal Processing
**Submitting author:** @lizliz (<a href="http://orcid.org/0000-0002-9459-9493">Elizabeth Munch</a>)
**Repository:** https://github.com/TeaspoonTDA/teaspoon
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** 1.5.7
