### 2024-08-26 #7145 @danielskatz
👋 @lizliz - thanks for your submission.  While we look for an editor, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.

Also, I wonder if the readme and/or the paper could use a little more information about this software for the general developer community (non-topology people)?
