### 2024-08-30 39c8a26
Updates with editor feedback
