+++ paper/paper.md
@@ -59,43 +59,59 @@ bibliography: paper.bib
 
 # Summary
 
-The field of topological data analysis (TDA) has risen in the last two decades to become a mature field of research providing a unique perspective on data analysis. 
-TDA consists of a suite of tools inspired by the field of Algebraic Topology which encodes shape and structure in data in a quantitative manner. 
-One particular subfield of work has focused on using TDA tools for the analysis of time series, colloquially known as Topological Signal Processing, or TSP. 
+The field of topological data analysis (TDA) has risen in the last two decades to become a mature field of research providing a unique perspective on data analysis.
+TDA consists of a suite of tools inspired by the field of Algebraic Topology which encodes shape and structure in data in a quantitative manner.
+One particular subfield of work has focused on using TDA tools for the analysis of time series, colloquially known as Topological Signal Processing, or TSP.
 The python package \texttt{teaspoon} has been built specifically to cater to the needs of researchers working in the field of TSP, with the added benefit that the code can be used for other forms of input data beyond signal.
-Recent work has largely focused on the use of persistent homology and its variants for this context, thus this has been the main tool utilized in \texttt{teaspoon}. 
-Full discussion of persistence is outside the scope of this brief paper, but we direct the interested reader to @Dey2021 and @Munch2017 for more information. 
 
-# Modules
+Recent work has largely focused on the use of persistent homology and its variants for this context, thus this has been the main tool utilized in \texttt{teaspoon}.
+While full discussion of the specifics of persistence is outside the scope of this brief paper, we give a brief introduction here and direct the interested reader to @Dey2021 and @Munch2017 for more information.
+Standard homology (see, e.g., @Hatcher) is a construction which builds a vector space for any input topological space $X$ and a given dimension $p$ of structure to be studied.
+Denoted $H_p(X)$, the $p=0$ dimensional homology encodes the structure of connected components; $p=1$ encodes loops; $p=2$ encodes voids; and higher dimensional versions exist without the interpretability of these lower dimensional versions.
+While homology is defined for a fixed topological space, persistent homology studies the changing homology for a changing topological space.
+Indeed, through a fundamental theorem of persistence [@CrawleyBoevey2015], we can use this sequence of vector spaces to determine when $p$-dimensional structures appear (or are born) and disappear (or die).
+The persistence diagram is a 2D scatter plot of points, each one giving the (birth, death) coordinates for a particular $p$-dimensional structure.
 
-Many topological signal processing projects can be fit into the pipeline shown in \autoref{fig:Pipeline}. 
-This pipeline aligns with the five submodules of \texttt{teaspoon}, which we describe in more detail in the following subsections.  
+![An example filtration of simplicial complexes and the corresponding persistence diagram. \label{fig:example}](figures/Filt_PD_Combined.png)
+
+As an example, in Figure \ref{fig:example} we have a sequence of topological spaces in which we wish to track how the homology is changing, along with the corresponding persistence diagram (below).
+Each red point in the persistence diagram at right represents a connected component (an $H_0$ feature), and each blue point represents a cycle (a $H_1$ feature).
+For example, three vertices from $K_0$ are connected by edges in $K_1$, thus there are two persistence points at $(0,1)$ in dimension 0 corresponding to the fact that three connected components have merged into one.
+Similarly, a cycle first appears in $K_2$ and is filled in at $K_4$, so there is a persistence point in dimension 1 at (2,4).
+
+In the following section, we show how persistence can be used in a pipeline for signal processing by aligning each of the steps with the modules of the \texttt{teaspoon} package. 
+
+# Package Modules
 
[…]
