+++ paper/paper.bib
@@ -113,15 +113,6 @@
   publisher = {American Physical Society (APS)},
 }
 
-@InProceedings{Myers2023b,
-  author    = {Myers, Audun D. and Kvinge, Henry and Emerson, Tegan},
-  booktitle = {Proceedings of the IEEE/CVF Conference on Computer Vision and Pattern Recognition (CVPR) Workshops},
-  title     = {TopFusion: Using Topological Feature Space for Fusion and Imputation in Multi-Modal Data},
-  year      = {2023},
-  month     = {June},
-  pages     = {600-609},
-}
-
  
 @Article{Guezel2022,
   author    = {Güzel, İsmail and Munch, Elizabeth and Khasawneh, Firas A.},
@@ -415,31 +406,6 @@
   readstatus   = {read},
 }
 
-@Article{Chevyrev2018,
-  author        = {Ilya Chevyrev and Vidit Nanda and Harald Oberhauser},
-  title         = {Persistence paths and signature features in topological data analysis},
-  year          = {2018},
-  month         = jun,
-  abstract      = {We introduce a new feature map for barcodes that arise in persistent homology computation. The main idea is to first realize each barcode as a path in a convenient vector space, and to then compute its path signature which takes values in the tensor algebra of that vector space. The composition of these two operations - barcode to path, path to tensor series - results in a feature map that has several desirable properties for statistical learning, such as universality and characteristicness, and achieves state-of-the-art results on common classification benchmarks.},
-  archiveprefix = {arXiv},
-  creationdate  = {2019-01-23T00:00:00},
-  eprint        = {http://arxiv.org/abs/1806.00381v1},
-  file          = {:Persistence/StatsAndPersistence/Machine Learning/Chevyrev2018.pdf:PDF;:Chevyrev2018 - Persistence Paths and Signature Features in Topological Data Analysis.pdf:PDF:http\://arxiv.org/pdf/1806.00381v1},
-  keywords      = {stat.ML, cs.LG, math.PR, math.ST, stat.TH, featurization},
-  primaryclass  = {stat.ML},
-}
-
-@InProceedings{Reininghaus2015,
-  author       = {Reininghaus, Jan and Huber, Stefan and Bauer, Ulrich and Kwitt, Roland},
-  booktitle    = {The IEEE Conference on Computer Vision and Pattern Recognition (CVPR)},
-  title        = {A Stable Multi-Scale Kernel for Topological Machine Learning},
-  year         = {2015},
-  month        = {June},
-  creationdate = {2017-01-30T00:00:00},
-  file         = {Reininghaus2015.pdf:Persistence/StatsAndPersistence/Reininghaus2015.pdf:PDF},
-  readstatus   = {read},
-}
-
 @Article{Dey2022,
   doi = {10.48550/arxiv.2204.11080},
   url = {https://arxiv.org/abs/2204.11080},
@@ -461,18 +427,6 @@
[…]
