### 2024-09-03 #7145 @hugoledoux
@lizliz I'm the editor, and things were a bit slow with the new academic year but now I'm looking for potential reviewers. Could you please look in our database of reviewers (https://reviewers.joss.theoj.org/) and suggest here 4-5 potential good ones, or someone else you think would be a good fit? Please remove their "@" from their github username to not ping them unnecessarily.
