### 2024-09-09 5ae0a19
PLandscapes functions updated with adding dtype parameters into np.array
