### 2024-09-10 9a53772
Incremented the version number and updated the documentation
