+++ docs/_sources/index.rst.txt
@@ -33,9 +33,9 @@ Modules
 .. image:: figures/tspPipeline.png
    :alt: The basic pipeline of a TSP project aligned with the modules of teaspoon.
 
-Many topological signal processing projects can be fit into the pipeline shown above; and this pipeline aligns with the five submodules of `teaspoon`. 
-We start with input time series data. In `teaspoon`, the `MakeData`_ module provides methods for creating synthetic data sets for testing purposes, including the `DynSysLib`_ submodule with an extensive array of dynamical systems. 
-The `ParameterSelection`_ modules provides tools for determining best parameters for transforming the time series into a delay coordinate (sometimes called Takens') embedding. This includes standard tools such as false nearest neighbors and mutual information function as well as newly developed tools. The `SP`_ module does the conversion from a time series to a topological input. This includes the standard delay coordinate embedding, as well as network representations of time series. This topological input can be converted to a topological signature, in most cases a persistence diagram, using the `TDA`_ module. Finally, these persistence diagrams can be used in a machine learning pipeline using the vectoriazation tools from the `ML`_ module.
+Many topological signal processing projects can be fit into the pipeline shown above; and this pipeline aligns with the five submodules of ``teaspoon``. 
+We start with input time series data. In ``teaspoon``, the `MakeData`_ module provides methods for creating synthetic data sets for testing purposes, including the `DynSysLib`_ submodule with an extensive array of dynamical systems. 
+The `ParameterSelection`_ modules provides tools for determining best parameters for transforming the time series into a delay coordinate (sometimes called Takens') embedding. This includes standard tools such as false nearest neighbors and mutual information function as well as newly developed tools. The Signal Processing (`SP`_) module does the conversion from a time series to a topological input. This includes the standard delay coordinate embedding, as well as network representations of time series. This topological input can be converted to a topological signature, in most cases a persistence diagram, using the `TDA`_ module. Finally, these persistence diagrams can be used in a machine learning pipeline using the vectorization tools from the `ML`_ module.
 
 .. _MakeData: https://teaspoontda.github.io/teaspoon/modules/MakeData/index.html
 .. _DynSysLib: https://teaspoontda.github.io/teaspoon/modules/MakeData/DynSysLib/index.html
