+++ docs/_sources/modules/MakeData/DynSysLib/delayed_flows.rst.txt
@@ -11,14 +11,3 @@ Main Functions
     :show-inheritance:
     :member-order: bysource
 
-
-Meta Function 
--------------
-
-This function is for being able to input into the original ``teaspoon.MakeData.DynSysLib.DynamicSystems`` function with default parameters.
-
-.. automodule:: teaspoon.MakeData.DynSysLib.delayed_flows
-    :members: delayed_flows
-    :undoc-members:
-    :show-inheritance:
-    :no-index:
