### 2024-10-06 TeaspoonTDA/teaspoon#136 @EduPH
[Joss review] Requirements
Hello, I am reviewing the package for your Joss submission. Adding a requirement file with python and the needed version of the packages would be nice. For example, I have the following error:

AttributeError: `np.infty` was removed in the NumPy 2.0 release. Use `np.inf` instead.

Thank you!
