### 2024-10-06 #7243 @EduPH
Hello @lizliz, thank you for such an excellent Python package. I am working on the review and added an issue in the repository about the needed version of the packages.
