+++ teaspoon/teaspoon/DAF/da_lib_AD.py
@@ -1,0 +1,165 @@
+"""
+    Library for topological data assimilation.
+
+"""
+import numpy                 as np
+
+import pandas                as pd
+import gudhi                 as gd
+from gudhi.tensorflow        import RipsLayer
+from matplotlib.gridspec     import GridSpec
+import os
+from gudhi.wasserstein       import wasserstein_distance
+from lr_forecast import get_forecast
+
+
+try:
+    import tensorflow as tf
+    from tensorflow import keras
+    os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3'
+
+except:
+    raise ImportError("TADA Requires tensorflow for optimization")
+
+def TADA(u_obs, window_size, model_parameters, n_epochs=1, train_len=4000, opt_params=[1e-8, 0.9, 1], window_number=1, stride=1, i=None):
+    '''
+        Compute optimal model weights using data assimilation and persistent homology.
+
+        Args:
+            X_train (array): Array of time series signals for training the initial model (D x N) D is the dimension, N is the number of time points
+            X_meas (array): Array of measurement time series signals (D x N) D is the dimension, N is the number of time points
+            n_epochs (int): Number of epochs for each assimilation window
+            save (bool): Boolean variable to control whether plots are saved as images
+            Dr (int): Reservoir dimension for generating the linear regression random feature map model
+            train_len (int):
+    '''
+
+    if window_number < window_size:
+        current_window_size = window_number + 1
+    else:
+        current_window_size = window_size
+
+    # Unpack forecast weights from random feature maps and linear regression
+    W_A, W_LR, W_in, b_in = model_parameters
+
+    W_LR = tf.constant(W_LR)
+
+    start = train_len
+    end = train_len + window_number + 1
+
+    X_meas = u_obs[:,start:end][:,-current_window_size:]
+    X_model = get_forecast(u_obs[:,train_len], W_A, W_in, b_in,forecast_len=end-train_len)[:,start:end]
+
+
+    initial_model = np.copy(X_model)
+
+    # Create tensorflow variable for weight matrix initialized to original LR weights
+    W = tf.Variable(initial_value=W_A, trainable=True, name="W", dtype=tf.float64)
+    
+
+    # Create rips layer for computing attaching edges 
+    rips_layer = RipsLayer(homology_dimensions=[0,1])
+
+    # Define optimizer and hyperparameters
+    l_rate = opt_params[0]
+    decay_rate = opt_params[1]
+    
+    if window_number > 1:
+        l_rate = l_rate * decay_rate**window_number
+
+    optimizer = keras.optimizers.Adam(learning_rate=l_rate)
+
+    # Compute initial persistence diagrams for measurement and model
+    target_pd0, target_pd1 = get_initial_pds(X_meas)
+    model_pd0, model_pd1 = get_initial_pds(X_model)
+    dgm0, dgm1 = model_pd0, model_pd1
+
+
+    # Initialize lists for loss values and persistence diagrams
+    losses = []
+    
+    #  Optimization Loop
+    for epoch in range(n_epochs):
+        if epoch != 0:
+            model_parameters = [W.numpy(), W_LR, W_in, b_in]
+        
+
+        start_pt = u_obs[:,train_len].reshape(-1,1)
+
+        # Optimization
+        with tf.GradientTape() as tape:
+            # Generate forecast with current model
+            
+            forecast = get_forecast(start_pt, W, W_in, b_in,forecast_len=window_number+1, auto_diff=True)
+        
[…]
