### 2024-11-01 935fae2
Clean up TADA function and add doc string
