### 2024-11-01 a83dd3f
Update docstrings for other functions in DA library.
