+++ doc_source/modules/DAF/index.rst
@@ -1,0 +1,12 @@
+.. _makedata:
+
+Data Assimilation and Forecasting Module Documentation
+=========================================================
+
+This module provides two methods for generating data for analysis. The first is through point cloud data generation for various shapes (e.g. sphere, annulus, etc.). The second module is the dynamic systems library. This library has approximately 60 dynamical systems available for simulation where the output is the time series of each variable of the dynamical system. These systems include flows, maps, and collected data.
+
+.. toctree::
+   :maxdepth: 1
+
+   Data Assimilation Module <DataAssimilation.rst>
+   Forecasting Module <Forecasting.rst>
