+++ docs/_sources/modules/DAF/index.rst.txt
@@ -1,4 +1,3 @@
-.. _makedata:
 
 Data Assimilation and Forecasting Module Documentation
 =========================================================
