### 2024-11-04 e79c864
Finishing forecasting documentation, updating dependency to add optional tensorflow dependency and perform all testing using install with TF.
