+++ teaspoon/teaspoon/DAF/data_assimilation.py
@@ -1,16 +1,9 @@
-"""
-    Library for topological data assimilation.
-
-"""
 import numpy                 as np
-
-import pandas                as pd
 import gudhi                 as gd
 from gudhi.tensorflow        import RipsLayer
-from matplotlib.gridspec     import GridSpec
 from gudhi.wasserstein       import wasserstein_distance
-from lr_forecast import get_forecast
-
+from teaspoon.DAF.forecasting import get_forecast
+from teaspoon.DAF.forecasting import forecast_time
 
 try:
     import tensorflow as tf
@@ -18,7 +11,7 @@ try:
 except:
     raise ImportError("TADA Requires tensorflow for optimization")
 
-def TADA(u_obs, window_size, model_parameters, n_epochs=1, train_len=4000, opt_params=[1e-8, 0.9, 1], window_number=1):
+def TADA(u_obs, window_size, model_parameters, n_epochs=1, train_len=4000, opt_params=[1e-5, 0.99], window_number=1):
     '''
         Compute optimal model weights using data assimilation and persistent homology. Heavily modified code from https://github.com/GUDHI/TDA-tutorial.git
 
@@ -32,7 +25,7 @@ def TADA(u_obs, window_size, model_parameters, n_epochs=1, train_len=4000, opt_p
             window_number (int): Current window number. Used to determine how many points to forecast into the future for the current window.
         
         Returns:
-            W (array): Optimal model weights using TADA algorithm.
+            (array): Optimal model weights using TADA algorithm.
     '''
 
     if window_number < window_size:
@@ -41,8 +34,10 @@ def TADA(u_obs, window_size, model_parameters, n_epochs=1, train_len=4000, opt_p
         current_window_size = window_size
 
     # Unpack forecast weights from random feature maps and linear regression
-    W_A, W_LR, W_in, b_in = model_parameters
+    if len(model_parameters) != 4:
+        raise ValueError("model_parameters must contain four items. W_A, W_LR, W_in, b_in")
 
+    W_A, W_LR, W_in, b_in = model_parameters
     W_LR = tf.constant(W_LR)
 
     start = train_len
@@ -61,32 +56,28 @@ def TADA(u_obs, window_size, model_parameters, n_epochs=1, train_len=4000, opt_p
     l_rate = opt_params[0]
     decay_rate = opt_params[1]
     
+    # Decay learning rate
     if window_number > 1:
         l_rate = l_rate * decay_rate**window_number
 
+    # Initialize optimizer
     optimizer = keras.optimizers.Adam(learning_rate=l_rate)
 
     # Compute initial persistence diagrams for measurement and model
     target_pd0, target_pd1 = get_initial_pds(X_meas)
     model_pd0, model_pd1 = get_initial_pds(X_model)
     dgm0, dgm1 = model_pd0, model_pd1
-
-
-    # Initialize lists for loss values and persistence diagrams
-    losses = []
     
-    #  Optimization Loop
+    #  Optimization Loop (allows for multiple epochs)
     for epoch in range(n_epochs):
         if epoch != 0:
             model_parameters = [W.numpy(), W_LR, W_in, b_in]
         
-
         start_pt = u_obs[:,train_len]
 
-        # Optimization
+        # Initialize gradient tape
         with tf.GradientTape() as tape:
             # Generate forecast with current model
-            
             forecast = get_forecast(start_pt, W, W_in, b_in,forecast_len=window_number+1, auto_diff=True)
         
             # Get persistence diagrams from current window forecast
@@ -94,14 +85,11 @@ def TADA(u_obs, window_size, model_parameters, n_epochs=1, train_len=4000, opt_p
             dgm0 = tf.cast(dgm[0][0], dtype=np.float64)
[…]
