+++ doc_source/modules/DAF/DataAssimilation.rst
@@ -1,4 +1,9 @@
 
 
-Data Assimilation Module Documentation
+Data Assimilation
 =========================================================
+
+This page gives a summary of the functions available in the data assimilation library.
+
+.. automodule:: teaspoon.DAF.data_assimilation
+    :members: 

[…]
