### 2024-11-05 437fe54
Update DAF unit test to use data file to mitigate floating point differences between operating systems
