+++ tests/test_DAF.py
@@ -11,7 +11,7 @@ class daf_module(unittest.TestCase):
         from teaspoon.DAF.forecasting import get_forecast
 
         # Set random seed
-        r_seed = 48824
+        r_seed = 123456
         np.random.seed(r_seed)
 
         # Set TADA parameters
@@ -23,9 +23,9 @@ class daf_module(unittest.TestCase):
         max_window_number = 10
 
         # Get training and validation data at random initial condition
-        ICs = list(np.random.normal(size=(3,1)).reshape(-1,))
-        t, ts = lorenz(L=500, fs=50, SampleSize=6001, parameters=[28,10.0,8.0/3.0])
-        ts = np.array(ts) 
+        import os
+        print(os.getcwd())
+        ts = np.load('./tests/DAF_lorenz_sim_data.npy') 
 
         # Get signal and noise amplitudes using signal-to-noise ratio
         a_sig = np.sqrt(np.mean(np.square(ts),axis=1))
@@ -66,9 +66,8 @@ class daf_module(unittest.TestCase):
         print("TADA Time: ", tada_time)
         print("LR Time: ", lr_time)
 
-        # Need to simulate lorenz with more accuracy for this to work.
-        # self.assertAlmostEqual(tada_time, 2.821, delta=0.001)
-        # self.assertAlmostEqual(lr_time, 1.875, delta=0.001)
+        self.assertAlmostEqual(tada_time, 3.658, delta=0.001)
+        self.assertAlmostEqual(lr_time, 3.058, delta=0.001)
 
 
 
