### 2024-11-05 d1490b1
Update pypi workflow files to install package with tensorflow
