### 2024-11-06 caf818e
Update publish_to_test_pypi.yml
Update test pypi workflow to run unit tests with installing optional dependencies. test_release uses the copy of the workflow file that is on the master branch. I am only proposing changes to the test_pypi file for now so nothing should be affected on the master branch.
