### 2024-11-25 #7243 @hugoledoux
👋 @lizliz what is the status with respect to the issues that were raised 4w ago by the reviewers?
