### 2024-11-26 dab6d1a
Add MANIFEST file to download EEG data with pip install
