+++ MANIFEST.in
@@ -1,0 +1,1 @@
+include teaspoon/teaspoon/MakeData/DynSysLib/data/EEG/*.txt
