### 2024-11-26 c81b475
Fix issue importing EEG data
Return type was a list with an array as its only element but now it is consistent with other systems in the library.
