+++ setup.py
@@ -9,4 +9,7 @@ if __name__ == "__main__":
     setup(packages = find_packages(),
     long_description = ld,
     dependency_links = ["git+https://github.com/shizuo-kaji/CubicalRipser_3dim.git"],
+    package_data={
+        'teaspoon.MakeData.DynSysLib.Data.EEG': ['Z093.txt'],
+    },
     include_package_data=True)
