### 2024-11-27 89dc2f1
Update MANIFEST.in
