### 2024-11-27 acd0176
Update MANIFEST.in
