+++ MANIFEST.in
@@ -1,1 +1,1 @@
-include teaspoon/MakeData/DynSysLib/data/EEG/*.txt
+include teaspoon/MakeData/DynSysLib/Data/EEG/*.txt
