### 2024-11-27 de8ce68
Update MANIFEST.in
