### 2024-11-27 63f2f6c
Update medical_data.py
