+++ MANIFEST.in
@@ -1,1 +1,1 @@
-include teaspoon/teaspoon/MakeData/DynSysLib/Data/EEG/*.txt
+include teaspoon/teaspoon/MakeData/DynSysLib/data/EEG/*.txt
