+++ .github/workflows/test_installs_data.yml
@@ -1,0 +1,276 @@
+name: Test Installs Data
+
+# Controls when the workflow will run
+on:
+
+
+  # Allows you to run this workflow manually from the Actions tab
+  workflow_dispatch:
+concurrency: release
+# A workflow run is made up of one or more jobs that can run sequentially or in parallel
+jobs:
+  build-39-ubuntu:
+    name: Build on Python 3.9 ubuntu
+    runs-on: ubuntu-latest
+    steps:
+      # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
+      - uses: actions/checkout@fix_medical_data
+        with:
+          ref: 'fix_medical_data'
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.9"
+      - name: Install pypa/build
+        run: >-
+          python -m
+          pip install
+          build
+          --user
+      - name: Install teaspoon for testing
+        run: pip install ".[full]"
+      - name: Run tests
+        run: make tests
+  build-39-mac:
+    name: Build on Python 3.9 mac
+    runs-on: macos-latest
+    steps:
+      # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
+      - uses: actions/checkout@master
+        with:
+          ref: 'test_release'
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.9"
+      - name: Install pypa/build
+        run: >-
+          python -m
+          pip install
+          build
+          --user
+      - name: Install teaspoon for testing
+        run: pip install ".[full]"
+      - name: Run tests
+        run: make tests
+  build-39-windows:
+    name: Build on Python 3.9 windows
+    runs-on: windows-latest
+    steps:
+      # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
+      - uses: actions/checkout@master
+        with:
+          ref: 'test_release'
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.9"
+      - name: Install pypa/build
+        run: >-
+          python -m
+          pip install
+          build
+          --user
+      - name: Install teaspoon for testing
+        run: pip install ".[full]"
+      - name: Run tests
+        run: make tests
+  build-310-ubuntu:
+    name: Build on Python 3.10 ubuntu
+    runs-on: ubuntu-latest
+    steps:
+      # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
+      - uses: actions/checkout@master
+        with:
+          ref: 'test_release'
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.10"
+      - name: Install pypa/build
+        run: >-
+          python -m
+          pip install
+          build
+          --user
+      - name: Install teaspoon for testing
+        run: pip install ".[full]"
+      - name: Run tests
+        run: make tests
+  build-310-mac:
+    name: Build on Python 3.10 mac
+    runs-on: macos-latest
+    steps:
+      # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
+      - uses: actions/checkout@master
+        with:
+          ref: 'test_release'
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.10"
+      - name: Install pypa/build
+        run: >-
+          python -m
+          pip install
+          build
+          --user
+      - name: Install teaspoon for testing
+        run: pip install ".[full]"
+      - name: Run tests
+        run: make tests
+  build-310-windows:
+    name: Build on Python 3.10 windows
+    runs-on: windows-latest
+    steps:
+      # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
+      - uses: actions/checkout@master
[…]
