### 2024-11-27 5e59deb
Test ubuntu install using fix medical data branch
