### 2024-11-27 ccd170f
Update test_installs.yml
