+++ .github/workflows/test_installs.yml
@@ -19,7 +19,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
-      - uses: actions/checkout@fix_medical_data
+      - uses: actions/checkout@master
         with:
           ref: 'fix_medical_data'
       - name: Set up Python
