### 2024-11-27 1d9b7b1
Update test_installs.yml
