+++ .github/workflows/test_installs.yml
@@ -21,7 +21,7 @@ jobs:
       # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
       - uses: actions/checkout@master
         with:
-          ref: 'fix_medical_data'
+          ref: 'test_release'
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
