### 2024-11-27 4cceb5c
Update pyproject.toml
