+++ teaspoon/teaspoon/MakeData/DynSysLib/__init__.py
@@ -1,1 +1,1 @@
-__all__ = ["data"]
+__all__ = ["Data"]
