### 2024-12-02 08d96a6
Update documentation to add TADA paper link.
