+++ .github/workflows/publish_to_test_pypi.yml
@@ -15,6 +15,7 @@ jobs:
     name: Build and publish python distributions to TestPyPI
     runs-on: ubuntu-latest
     concurrency: release
+    if: github.event.workflow_run.head_branch == 'test_release'  # Only run if the first workflow was on the 'test_release' branch
     steps:
       # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
       - uses: actions/checkout@master
