### 2024-12-04 542e280
Clean up parameter selection documentation
