### 2025-01-08 4281971
Updates for the joss review
