+++ requirements.txt
@@ -1,0 +1,72 @@
+pip
+absl-py
+astunparse
+certifi
+charset-normalizer
+contourpy
+cycler>=0.11.0
+Cython
+ddeint
+Deprecated
+eagerpy
+flatbuffers
+fonttools>=4.34.4
+gast
+google-pasta
+grpcio
+gudhi>=3.10.1
+h5py
+hopcroftkarp
+idna
+importlib_resources
+joblib
+keras
+kiwisolver
+libclang
+llvmlite
+Markdown
+markdown-it-py
+MarkupSafe
+matplotlib>=3.5.2
+mdurl
+ml-dtypes
+mpmath
+namex
+networkx>=3.1
+numba
+numpy
+opt_einsum
+optree
+packaging
+pandas>=1.4.1
+persim
+pillow
+platformdirs
+pooch
+POT>=0.8.2
+protobuf
+pyentrp
+Pygments
+pyparsing
+python-dateutil
+pytz
+requests
+rich
+ripser>=0.6.4
+scikit-learn>=1.1.1
+scipy>=1.8.1
+six>=1.16.0
+sortedcontainers
+sympy
+teaspoon>=1.4
+tensorboard
+tensorboard-data-server
+tensorflow
+tensorflow-io-gcs-filesystem
+termcolor
+threadpoolctl
+typing_extensions
+tzdata
+urllib3
+Werkzeug
+wrapt
