### 2025-01-24 #7243 @EduPH
Dear @lizliz, I have finished the review. Thank you for your nice work and the great contribution to the TDA community.

@hugoledoux, all done from my side.
