### 2025-01-25 8d7bbdc
Add RAFDA to DA library
