### 2025-03-05 #7243 @hugoledoux
@lizliz I have fixed a few things in your paper, there are 2 PR, please merge

when I have the go from @EduPH we'll wrap up the acceptance
