### 2025-03-05 #7243 @EduPH
@hugoledoux Yes, it's because there are no original results and no animal abuse. Sorry for the delay. I changed institutions recently, and tracking emails is a nightmare.

Again, thank @lizliz for the contribution. Teaspoon is a very useful library.
