### 2025-03-05 #7243 @hugoledoux
@EduPH great if you could tick them, for admin purposes, it would be great
