### 2025-03-07 #7243 @lizliz
Hi Hugo!

Thanks for the feedback. I'm working on getting the rest of those todos taken care of.

In the mean time, I'm hoping you can help me figure out what changed in the last few months that my compiler on github for the paper no longer works. Do you have any advice on how to fix this? https://github.com/TeaspoonTDA/teaspoon/actions/runs/13723177055/job/38383237124

Thanks in advance!
