### 2025-03-07 #7243 @hugoledoux
@editoralbot generate pdf
