### 2025-03-07 #7243 @hugoledoux
it compiles from here, see above.

I guess https://github.com/TeaspoonTDA/teaspoon/blob/joss-paper/.github/workflows/draft-pdf.yml#L17 should "@v4", otherwise no idea
