### 2025-03-07 #7243 @hugoledoux
sorry the names on the paper and on zenodo need to be **exactly** the same, including the middle names.

can you update the paper please?
