### 2025-03-07 #7243 @lizliz
Better yet, I was able to fix the metadata on zenodo. Check me, but I think they all match now.
