### 2025-03-11 #7243 @hugoledoux
@yossibokorbleile @EduPH thanks for your help, without our reviewers JOSS wouldn't exist!

@lizliz it's now accepted, but one final check needs to be done by one of the editor-in-chief, they might have some minor changes
