### 2025-03-13 a84b6f9
partial changes in paper.bib
