+++ paper/paper.bib
@@ -349,7 +349,7 @@
 @InProceedings{BinMasood2019,
   author    = {Bin Masood, Talha and Budin, Joseph and Falk, Martin and Favelier, Guillaume and Garth, Christoph and Gueunet, Charles and Guillou, Pierre and Hofmann, Lutz and Hristov, Petar and Kamakshidasan, Adhitya and Kappe, Christopher and Klacansky, Pavol and Laurin, Patrick and Levine, Joshua and Lukasczyk, Jonas and Sakurai, Daisuke and Soler, Maxime and Steneteg, Peter and Tierny, Julien and Usher, Will and Vidal, Jules and Wozniak, Michal},
   booktitle = {{TopoInVis}},
-  title     = {{An Overview of the Topology ToolKit}},
+  title     = {An Overview of the {Topology ToolKit}},
   year      = {2019},
 }
 
