### 2025-03-13 81e6030
more small changes in paper.bib
