+++ paper/paper.bib
@@ -144,7 +144,7 @@
 
 @Article{Jones2023,
   author        = {Jones, Benjamin and Wei, Guowei},
-  title         = {Persistent Directed Flag Laplacian},
+  title         = {Persistent Directed Flag {L}aplacian},
   year          = {2023},
   month         = dec,
   abstract      = {Topological data analysis (TDA) has had enormous success in science and engineering in the past decade. Persistent topological Laplacians (PTLs) overcome some limitations of persistent homology, a key technique in TDA, and provide substantial insight to the behavior of various geometric and topological objects. This work extends PTLs to directed flag complexes, which are an exciting generalization to flag complexes, also known as clique complexes, that arise naturally in many situations. We introduce the directed flag Laplacian and show that the proposed persistent directed flag Laplacian (PDFL) is a distinct way of analyzing these flag complexes. Example calculations are provided to demonstrate the potential of the proposed PDFL in real world applications.},
@@ -202,7 +202,7 @@
 
 @Article{Elchesen2022,
   author        = {Elchesen, Alex and Hartsock, Iryna and Perea, Jose A. and Rask, Tatum},
-  title         = {Learning on Persistence Diagrams as Radon Measures},
+  title         = {Learning on Persistence Diagrams as {R}adon Measures},
   year          = {2022},
   month         = dec,
   abstract      = {Persistence diagrams are common descriptors of the topological structure of data appearing in various classification and regression tasks. They can be generalized to Radon measures supported on the birth-death plane and endowed with an optimal transport distance. Examples of such measures are expectations of probability distributions on the space of persistence diagrams. In this paper, we develop methods for approximating continuous functions on the space of Radon measures supported on the birth-death plane, as well as their utilization in supervised learning tasks. Indeed, we show that any continuous function defined on a compact subset of the space of such measures (e.g., a classifier or regressor) can be approximated arbitrarily well by polynomial combinations of features computed using a continuous compactly supported function on the birth-death plane (a template). We provide insights into the structure of relatively compact subsets of the space of Radon measures, and test our approximation methodology on various data sets and supervised learning tasks.},
@@ -284,7 +284,7 @@
 
 @Misc{Saul2019,
   author = {Nathaniel Saul and Chris Tralie},
-  title  = {Scikit-TDA: Topological Data Analysis for Python},
+  title  = {Scikit-TDA: Topological Data Analysis for {P}ython},
   year   = {2019},
   doi    = {10.5281/zenodo.2533369},
   url    = {https://doi.org/10.5281/zenodo.2533369},
@@ -474,7 +474,7 @@
 @Conference{Myers2020,
   author    = {Myers, Audun D. and Melih Yesilli and Sarah Tymochko and Firas Khasawneh and Elizabeth Munch},
   booktitle = {Topological Data Analysis and Beyond Workshop at NeurIPS 2020},
-  title     = {Teaspoon: A comprehensive python package for topological signal processing},
+  title     = {Teaspoon: A comprehensive {P}ython package for topological signal processing},
   year      = {2020},
   image     = {teaspoon.png},
   keywords  = {conference},
