### 2025-03-13 #7243 @danielskatz
I've suggested a number of small changes in https://github.com/TeaspoonTDA/teaspoon/pull/169 - please merge this, or let me know what you disagree with, then we can continue to publishing the work.
