### 2025-03-14 #7243 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
