### 2025-03-14 #7243 @danielskatz
Congratulations to @lizliz (Elizabeth Munch) and co-authors on your JOSS publication!!

JOSS completely depends on volunteers and we couldn't do this work without you.
